Fixed issue with dmsdk headers including internal headers
