+++ engine/gui/src/gui.h
@@ -44,9 +44,6 @@
  */
 namespace dmGui
 {
-    typedef struct Context* HContext;
-    typedef struct Script* HScript;
-
     /**
      * Default layout id
      */
