+++ engine/sdk/src/test/test_include.cpp
@@ -1,0 +1,6 @@
+#include <dmsdk/sdk.h>
+#include <{custom_include}>
+int main(int argc, char **argv)
+{
+    return 0;
+}
